Added a property for physics suspension to the GameObject subclass. Added option to clear user rotation property by setting it to None in the GameObject instance. Changed some of the demonstration applications to use the rest Application class.
Added an option to calibrate on-the-fly to the textutils text width measurement. Couldn't get a reliable fixed calibration ratio.
Added textutils initialisation and a method into the base Application class.
Added a GameText custom subclass that does the same job as the GameObject subclass.

Added more persistent storage capabilities to the unittest Application, which runs unit tests in Blender.
Added a per-test method status message that is displayed in BGE.
Changed the test methods to be compatible with the thread Application class.

Added setattr to InterceptProperty.

Consistency improvements in comments and such like.