WiP commit.
Changed RestAnimation.path to valuePath. Added RestAnimation.subjectPath and subject to hold the path and reference to the animated object.
Stopped using Animation.userData to track completed animations for physics resumption, and Rotation reset.

Added expectedFailure support to ThreadTestResult but didn't use it in the end.

Moved replacement of completed animation with None down into the rest Application class. Added assertions to the relevant test case.