Added to the GameObject class a method to set parent that treats None as removing the parent. Changed the Cursor class to utilise it.
Added an exception when attempting to end an object whose tether has children.

Moved the GameObject rotation clearing code to the physics setter from the beingAnimated setter.
Changed the demonstration user interface to rely on that, instead of nulling every object before patching it during build.

Changed the toBuild array in the user interface to be a parameter to the build_ methods, instead of being an instance property.

Fixed the user interface clear monitor button to delete the progress message too.

Changed the build so that trackBuild:false means the cursor subject doesn't get changed at all.

Updated backlog. Some late deletions for things already done but mostly new things.